STAF-36: fix errors
