--- src/core/components/Page/index.ts
@@ -1,1 +1,0 @@
-export { Page } from './Page';
