+++ src/core/components/PageContent/PageContent.css
@@ -1,4 +1,4 @@
-.page {
+.page-content {
   position: relative;
   overflow: auto;
   width: 100%;
