+++ src/core/components/PageContent/index.ts
@@ -1,0 +1,1 @@
+export { PageContent } from './PageContent';
