+++ src/core/components/ProtectedPageLayout/ProtectedPageLayout.css
@@ -1,5 +1,4 @@
-.root-layout-container {
-  display: grid;
+.protected-page-layout {
   position: fixed;
   overflow: auto;
   width: 100%;
@@ -10,7 +9,7 @@
   left: 0;
 }
 
-.root-layout-header {
+.protected-page-layout__header {
   position: fixed;
   top: 0;
   width: 100%;
@@ -26,7 +25,7 @@
   padding: 0 3rem;
 }
 
-.root-content-container {
+.protected-page-layout__content-container {
   position: relative;
   top: 3.75rem;
   display: flex;
@@ -35,7 +34,7 @@
   align-content: center;
 }
 
-.root-layout-footer {
+.protected-page-layout__footer {
   position: fixed;
   bottom: 0;
   display: flex;
