+++ src/core/components/ProtectedPageLayout/index.ts
@@ -1,0 +1,1 @@
+export { ProtectedPageLayout as RootNavLayout } from './ProtectedPageLayout';
