--- src/core/components/RootNavLayout/index.ts
@@ -1,1 +1,0 @@
-export { RootNavLayout } from './RootNavLayout';
