+++ src/shift/components/ShiftsPage.tsx
@@ -1,13 +1,13 @@
 import { Link } from 'react-router-dom';
-import { Page } from '../../core/components/Page';
+import { PageContent } from '../../core/components/PageContent';
 import { ProtectedRoutesEnum } from '../../config/routes/routes';
 import { ShiftListContainer } from './ShiftListContainer';
 
 export const ShiftsPage = (): JSX.Element => {
   return (
-    <Page classes="shifts-page">
+    <PageContent classes="shifts-page">
       <Link to={ProtectedRoutesEnum.shiftNew}>New Shift</Link>
       <ShiftListContainer />
-    </Page>
+    </PageContent>
   );
 };
