+++ src/team/components/TeamPage.tsx
@@ -1,13 +1,13 @@
 import { Link } from 'react-router-dom';
-import { Page } from '../../core/components/Page';
+import { PageContent } from '../../core/components/PageContent';
 import { ProtectedRoutesEnum } from '../../config/routes/routes';
 import { TeamListContainer } from './TeamListContainer';
 
 export const TeamPage = (): JSX.Element => {
   return (
-    <Page classes="new-team-page">
+    <PageContent classes="new-team-page">
       <Link to={ProtectedRoutesEnum.teamNew}>New Team</Link>
       <TeamListContainer />
-    </Page>
+    </PageContent>
   );
 };
