+++ src/timeSheet/TimeSheetsPage.tsx
@@ -1,11 +1,11 @@
 import { Link } from 'react-router-dom';
-import { Page } from '../core/components/Page';
+import { PageContent } from '../core/components/PageContent';
 import { ProtectedRoutesEnum } from '../config/routes/routes';
 
 export const TimeSheetsPage = (): JSX.Element => {
   return (
-    <Page classes="time-sheets-page">
+    <PageContent classes="time-sheets-page">
       <Link to={ProtectedRoutesEnum.timeSheets}>Time Sheets</Link>
-    </Page>
+    </PageContent>
   );
 };
